// src/api/src/resources/interface/job.queue.interface.ts
export enum JobQueueCommand {
    PRODUCT_INSERT = 'INSERT_PRODUCT',
    PRODUCT_UPDATE = 'UPDATE_PRODUCT',
    PRODUCT_DELETE = 'DELETE_PRODUCT',

    ITEM_PAUSE = 'PAUSE_ITEM',

    ITEMS_DOWNLOAD = 'DOWNLOAD_ITEMS',
    ITEM_DOWNLOAD = 'DOWNLOAD_ITEM',
    ITEMS_UPDATE = 'UPDATE_ITEMS',
    ITEM_UPDATE = 'UPDATE_ITEM',
    ITEM_LINK = 'LINK_ITEM',
    ITEM_UNLINK = 'UNLINK_ITEM',
    ITEM_VARIATION_LINK = 'LINK_ITEM_VARIATION',
    ITEM_VARIATION_UNLINK = 'UNLINK_ITEM_VARIATION',

    NOTIFICATION_ORDERS = 'ORDER_NOTIFICATION',
    NOTIFICATION_ORDERS_V2 = 'ORDER_NOTIFICATION',
    NOTIFICATION_ITEMS = 'ITEM_NOTIFICATION',
    NOTIFICATION_PAYMENTS = 'PAYMENT_NOTIFICATION',
    NOTIFICATION_CREATED_ORDERS = 'CREATED_ORDERS_NOTIFICATION',
    NOTIFICATION_QUESTIONS = 'QUESTIONS_NOTIFICATION',
    NOTIFICATION_PICTURES = 'PICTURES_NOTIFICATION',
    NOTIFICATION_MESSAGES = 'MESSAGING_NOTIFICATION',
    NOTIFICATION_SHIPMENTS = 'SHIPMENTS_NOTIFICATION',
    NOTIFICATION_QUOTATIONS = 'QUOTATIONS_NOTIFICATION',
    NOTIFICATION_INVOICES = 'INVOICES_NOTIFICATION',
    NOTIFICATION_CLAIMS = 'CLAIMS_NOTIFICATION',
    NOTIFICATION_ITEM_COMPETITION = 'ITEM_COMPETITION_NOTIFICATION',
}

export interface JobQueueInterface<T> {
    item: T;
    command: JobQueueCommand;
}